LayoutTests/http/tests/inspector/console-xhr-logging-async.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="console-test.js"></script>
<script src="network-test.js"></script>
<script>
function makeRequest()
{
    doXHR("GET", "resources/xhr-exists.html", true, allXHRsLoaded);
}

function allXHRsLoaded()
{
    console.log("Done.");
}

function test()
{
    step1();

    function step1()
    {
        function callback()
        {
            InspectorTest.evaluateInPage("makeRequest()");
        }

        InspectorTest.addConsoleSniffer(step2);
        ConsoleAgent.setMonitoringXHREnabled(true, callback);
    }

    function step2()
    {
        function callback()
        {
            InspectorTest.evaluateInPage("makeRequest()");
        }

        InspectorTest.addConsoleSniffer(step3);
        ConsoleAgent.setMonitoringXHREnabled(false, callback);
    }

    function step3()
    {
        function finish()
        {
            InspectorTest.dumpConsoleMessages();
            InspectorTest.completeTest();
        }
        InspectorTest.runAfterPendingDispatches(finish);
    }
}

</script>
</head>

<body onload="runTest()">
<p>Tests that XMLHttpRequest Logging works when Enabled and doesn't show logs when Disabled for asynchronous XHRs.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=79229">Bug 79229</a>
</body>
</html>
